Fix Dockerfile for Azure Container Registry builds
As commented in #3 by @horihiro:

> According to the implementation of vscode-docker, it seems that the
> Build image in Azure command uses the root of the workspace as a root
> directory for container-building even if Dockerfile is in a
> sub-directry.

This commit moves the Dockerfile to the root of the repository. This
file was updated to reflect this move, removed duplicate work from
`COPY` commands and ensure the same file work consistently for both
local builds and for Azure Container Registry builds.

Fixes #3

Also addresses MicrosoftDocs/azure-docs#101118

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>